BFC
This guys  a doozy, I've refactored the database IO, switching out my old DBO class for the new SQLAdapter which is much nicer to use, more conscise, has more functionality and is easier to extend the functionality of.
Also re-shuffled the junble of files into folders so the whole thing is better organised.
Finally added the backend functionality to add a new product to a package.